+++ lib/backend/placeFormHandler.js
@@ -4,7 +4,9 @@ const utils = require.main.require('./src/utils');
 const meta = require.main.require('./src/meta');
 const db = require.main.require('./src/database');
 const file = require.main.require('./src/file');
+const posts = require.main.require('./src/posts');
 const topics = require.main.require('./src/topics');
+const privileges = require.main.require('./src/privileges');
 const { uploadImage, resizeImage } = require.main.require('./src/image');
 const inputValidator = require('./inputValidator');
 
@@ -154,13 +156,51 @@ async function handleAddPlaceRequest(req, res, helpers) {
     }
 }
 
+
+async function editTopic(tid, topicData, uid) {
+	return new Promise(async (resolve, reject) => {
+	  try {
+		const topic = await topics.getTopicData(tid);
+		
+		
+		if (!topic?.mainPid) {
+		  resolve({ error: 'Topic not found' });
+		  return;
+		}
+  
+		const canEdit = await privileges.posts.canEdit(topic.mainPid, uid);
+		if (!canEdit) {
+		  resolve({ error: 'You are not allowed to edit this post' });
+		  return;
+		}
+  
+		
+
+		// topic.title = topicData.title;
+		// topic.content = topicData.content; 
+		// topic.tags = topicData.tags;
+
+		await topics.setTopicFields(tid, topicData)
+		await posts.edit({
+			uid: uid,
+			pid: topic.mainPid,
+			content: topicData.content
+		});
+
+        resolve(topic);
+	  } catch (error) {
+		resolve({ error });
+	  }
+	});
+}
+
 module.exports = handleAddPlaceRequest;
 
 
 // const topicModifier = require('./lib/backend/topicModifier');
-// const privileges = require.main.require('./src/privileges');
+// 
 // const routeHelpers = require.main.require('./src/routes/helpers');
-// const posts = require.main.require('./src/posts');
+// 
 // const nconf = require.main.require('nconf');
 // const cache = require.main.require('./src/cache');
 // const plugins = require.main.require('./src/plugins');
